Make nrf52 run correctly
I had to fixup the way that libraries were linked in; some of the
assembly files in the core libraries can't be linked into the same
executable when referenced as object files, but can if we stick
them in an `ar` archive.

Switching to library archives meant making sure that we list the
libraries after the object files.  What object files?  We have
to make sure that the main firmware project uses object files
explicitly too.

Since the variant and the core libraries are inter-dependent,
the sources for both need to included in the same library
archive.

But: now each of the boards referenced in the blinky project can
successfully blink an LED using tasks created via FreeRTOS APIs.