Add a string library
This is a little bit NIH, but the ideal here is that we can use
generic programming techniques to work with strings and not care
so much about flash string helpers.

In addition, the FixedString class contained in here doesn't
require any dynamic allocation.

The string implementation isn't complete; beyond initializing,
appending and equality comparisons, most of the usual string
functions are currently missing.