Allow pulling in keyboard layout editor data files
There's some small code generation going on here at the moment.
I want to expand this so that we provide a KLE file with labels
like `k00`, `k11` to define the physical position to matrix mapping,
and/or find a way to provide that mapping in the info.py?

I've pulled in some code from my other utils repo; thinking about
rendering an SVG as part of this, but haven't hooked up the code
for this bit.