fixup scan code and matrix handling
Was using the wrong math to compute the scan code and thus leaving out
portions of the matrix.  This was compounded by an off-by-one error
when iterating the rows and cols.

Added in support for translating the modifier keys to modifier flags,
so we now have a fully functioning basic keyboard.

Next steps: layers, consumers key codes.