Add basic ergodox (ez) support
This adds a scanner that knows how to read the split keyboard
present in the ergodox ez.

It also pulls in an adjusted version of the LUFA TWI code; it
needed to be modified to wait for the stop transmission to be
completed before proceeding with other communications on the I2C
bus.  Without this, the left hand would not be read correctly.