allow sleeping the scanner and bluetooth tasks
When the dispatcher task detects no change in the matrix for an
extended period, it calls an application provided idleSleep()
method.

For the flutterby this switches the expander into interrupt driven
mode and then blocks the scanner task until an interrupt is generated.

In addition to this, the idleSleep method instructs the ble task
to power down the device.